requirements.txt
praw==7.7.1
prawcore==2.3.0
pydantic==1.10.8
PyYAML==6.0
superagi-tools==1.0.6
